Refactor LoadsTools to integrate load retrieval methods directly; enhance error handling with JSON-RPC error codes and improve context propagation
