+++ spring-loaddev-mcp-server/src/main/java/ca/zhoozhoo/loaddev/mcp/tools/LoadsTools.java
@@ -14,36 +14,61 @@
  */
 package ca.zhoozhoo.loaddev.mcp.tools;
 
+import static io.modelcontextprotocol.spec.McpSchema.ErrorCodes.INTERNAL_ERROR;
+import static io.modelcontextprotocol.spec.McpSchema.ErrorCodes.INVALID_PARAMS;
+import static io.modelcontextprotocol.spec.McpSchema.ErrorCodes.INVALID_REQUEST;
+import static org.springframework.http.HttpStatus.NOT_FOUND;
+import static org.springframework.http.HttpStatus.UNAUTHORIZED;
 import static reactor.core.publisher.Mono.just;
+import static java.lang.String.format;
 
+import java.time.Duration;
 import java.util.List;
 
 import org.springframework.ai.chat.model.ToolContext;
 import org.springframework.ai.tool.annotation.Tool;
 import org.springframework.ai.tool.annotation.ToolParam;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.cloud.client.discovery.DiscoveryClient;
+import org.springframework.security.core.context.ReactiveSecurityContextHolder;
+import org.springframework.security.core.context.SecurityContext;
+import org.springframework.security.oauth2.jwt.Jwt;
 import org.springframework.stereotype.Service;
+import org.springframework.web.reactive.function.client.WebClient;
+import org.springframework.web.reactive.function.client.WebClientResponseException;
 
 import ca.zhoozhoo.loaddev.mcp.config.McpToolRegistrationConfig.ReactiveContextHolder;
 import ca.zhoozhoo.loaddev.mcp.dto.GroupStatisticsDto;
 import ca.zhoozhoo.loaddev.mcp.dto.LoadDto;
-import ca.zhoozhoo.loaddev.mcp.service.LoadsService;
 import io.modelcontextprotocol.spec.McpError;
-import io.modelcontextprotocol.spec.McpSchema;
+import io.modelcontextprotocol.spec.McpSchema.JSONRPCResponse.JSONRPCError;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
+import reactor.core.publisher.Flux;
+import reactor.core.publisher.Mono;
 import reactor.util.context.ContextView;
+import reactor.util.retry.Retry;
 
 @Service
 @Log4j2
 @RequiredArgsConstructor
 public class LoadsTools {
-    private final LoadsService loadsService;
+
+    @Autowired
+    private WebClient webClient;
+
+    @Autowired
+    private DiscoveryClient discoveryClient;
+
+    @Value("${service.loads.name:loads-service}")
+    private String loadsServiceName;
 
     /**
      * Retrieves a specific load by its ID.
      * Ensures reactive context propagation and proper error handling.
      *
-     * @param id ID of the load to retrieve
+     * @param id      ID of the load to retrieve
      * @param context Tool execution context
      * @return The requested load
      * @throws McpError with INTERNAL_ERROR code if reactive context is missing
@@ -52,31 +77,72 @@ public class LoadsTools {
      */
     @Tool(description = "Find a specific load by its unique identifier", name = "getLoadById")
     public LoadDto getLoadById(
-            @ToolParam(description = "Numeric ID of the load to retrieve") Long id,
+            @ToolParam(description = "Numeric ID of the load to retrieve", required = true) Long id,
             ToolContext context) {
         log.debug("Retrieving load with ID: {}", id);
-        ContextView reactiveContext = getReactiveContext();
+
+        if (id == null || id <= 0) {
+            throw new McpError(new JSONRPCError(
+                    INVALID_PARAMS,
+                    "Load ID must be a positive number",
+                    null));
+        }
+
+        var reactiveContext = getReactiveContext();
 
         try {
             return just(id)
-                    .flatMap(loadId -> loadsService.getLoadById(loadId))
+                    .flatMap(loadId -> getLoadById(loadId))
                     .contextWrite(ctx -> ctx.putAll(reactiveContext))
                     .doOnSuccess(load -> log.debug("Successfully retrieved load: {}", load))
-                    .doOnError(IllegalArgumentException.class, 
-                        e -> log.debug("Load not found with ID {}: {}", id, e.getMessage()))
-                    .doOnError(SecurityException.class, 
-                        e -> log.error("Authentication error retrieving load {}: {}", id, e.getMessage()))
+                    .doOnError(IllegalArgumentException.class,
+                            e -> log.debug("Load not found with ID {}: {}", id, e.getMessage()))
+                    .doOnError(SecurityException.class,
+                            e -> log.error("Authentication error retrieving load {}: {}", id, e.getMessage()))
                     .doOnError(e -> log.error("Error retrieving load {}: {}", id, e.getMessage()))
                     .block();
         } catch (Exception e) {
             if (e.getCause() instanceof IllegalArgumentException) {
-                throw new McpError(new McpSchema.JSONRPCResponse.JSONRPCError(
-                    McpSchema.ErrorCodes.INVALID_PARAMS, 
-                    "Load not found: " + e.getCause().getMessage(), 
-                    null));
+                throw new McpError(new JSONRPCError(
+                        INVALID_PARAMS,
+                        "Load not found: " + e.getCause().getMessage(),
+                        null));
             }
             throw handleException("Failed to retrieve load", e);
         }
+    }
+
+    /**
+     * Retrieves a single load by its ID from the loads service.
+     * 
+     * @param id The ID of the load to retrieve
+     * @return A Mono of LoadDto object
+     */
+    private Mono<LoadDto> getLoadById(Long id) {
+        return ReactiveSecurityContextHolder.getContext()
+                .map(SecurityContext::getAuthentication)
+                .flatMap(auth -> {
+                    var instances = discoveryClient.getInstances(loadsServiceName);
+                    if (instances == null || instances.isEmpty()) {
+                        return Mono.error(new IllegalStateException(
+                                String.format("Service %s not found in discovery", loadsServiceName)));
+                    }
+
+                    return webClient
+                            .get()
+                            .uri(instances.get(0).getUri().toString() + "/loads/" + id)
+                            .headers(h -> h.setBearerAuth(((Jwt) auth.getCredentials()).getTokenValue()))
+                            .retrieve()
+                            .bodyToMono(LoadDto.class)
+                            .onErrorMap(WebClientResponseException.class,
+                                    e -> e.getStatusCode() == UNAUTHORIZED
+                                            ? new SecurityException("Authentication failed", e)
+                                            : e.getStatusCode() == NOT_FOUND
+                                                    ? new IllegalArgumentException("Load not found with ID: " + id)
+                                                    : e)
+                            .retryWhen(Retry.backoff(3, Duration.ofSeconds(1))
+                                    .filter(throwable -> !(throwable instanceof SecurityException)));
+                });
     }
 
     /**
@@ -91,33 +157,63 @@ public class LoadsTools {
     @Tool(description = "Retrieve all available loads in the system", name = "getLoads")
     public List<LoadDto> getLoads(ToolContext context) {
         log.debug("Retrieving all loads");
-        ContextView reactiveContext = getReactiveContext();
+        var reactiveContext = getReactiveContext();
 
         try {
-            return loadsService.getLoads()
+            return getLoads()
                     .contextWrite(ctx -> ctx.putAll(reactiveContext))
                     .collectList()
                     .doOnSuccess(list -> log.debug("Successfully retrieved {} loads", list.size()))
-                    .doOnError(SecurityException.class, 
-                        e -> log.error("Authentication error retrieving loads: {}", e.getMessage()))
+                    .doOnError(SecurityException.class,
+                            e -> log.error("Authentication error retrieving loads: {}", e.getMessage()))
                     .doOnError(e -> log.error("Error retrieving loads: {}", e.getMessage()))
                     .block();
         } catch (Exception e) {
             if (e.getCause() instanceof SecurityException) {
-                throw new McpError(new McpSchema.JSONRPCResponse.JSONRPCError(
-                    McpSchema.ErrorCodes.INVALID_REQUEST,
-                    "Authentication failed while retrieving loads: " + e.getMessage(),
-                    null));
+                throw new McpError(new JSONRPCError(
+                        INVALID_REQUEST,
+                        "Authentication failed while retrieving loads: " + e.getMessage(),
+                        null));
             }
             throw handleException("Failed to retrieve loads", e);
         }
     }
 
     /**
+     * Retrieves all loads from the loads service.
+     * 
+     * @return A Flux of LoadDto objects
+     */
+    private Flux<LoadDto> getLoads() {
+        return ReactiveSecurityContextHolder.getContext()
+                .map(SecurityContext::getAuthentication)
+                .flatMapMany(auth -> {
+                    var instances = discoveryClient.getInstances(loadsServiceName);
+                    if (instances == null || instances.isEmpty()) {
+                        return Flux.error(new IllegalStateException(
+                                String.format("Service %s not found in discovery", loadsServiceName)));
+                    }
+
+                    return webClient
+                            .get()
+                            .uri(instances.get(0).getUri().toString() + "/loads")
+                            .headers(h -> h.setBearerAuth(((Jwt) auth.getCredentials()).getTokenValue()))
+                            .retrieve()
+                            .bodyToFlux(LoadDto.class)
+                            .onErrorMap(WebClientResponseException.class,
+                                    e -> e.getStatusCode() == UNAUTHORIZED
+                                            ? new SecurityException("Authentication failed", e)
+                                            : e)
+                            .retryWhen(Retry.backoff(3, Duration.ofSeconds(1))
+                                    .filter(throwable -> !(throwable instanceof SecurityException)));
+                });
+    }
+
+    /**
      * Retrieves statistics for a specific load.
      * Ensures reactive context propagation and proper error handling.
      *
-     * @param id ID of the load to retrieve statistics for
+     * @param id      ID of the load to retrieve statistics for
      * @param context Tool execution context
      * @return List of group statistics for the load
      * @throws McpError with INTERNAL_ERROR code if reactive context is missing
@@ -126,31 +222,53 @@ public class LoadsTools {
      */
     @Tool(description = "Get statistics for a specific load", name = "getLoadStatistics")
     public List<GroupStatisticsDto> getLoadStatistics(
-            @ToolParam(description = "Numeric ID of the load") Long id,
-            ToolContext context) {
+            @ToolParam(description = "Numeric ID of the load", required = true) Long id, ToolContext context) {
         log.debug("Retrieving statistics for load ID: {}", id);
-        ContextView reactiveContext = getReactiveContext();
 
-        try {
-            return just(id)
-                    .flatMapMany(loadId -> loadsService.getLoadStatistics(loadId))
-                    .contextWrite(ctx -> ctx.putAll(reactiveContext))
-                    .collectList()
-                    .doOnSuccess(stats -> log.debug("Retrieved {} statistics for load {}", stats.size(), id))
-                    .doOnError(IllegalArgumentException.class, 
-                        e -> log.debug("Load not found with ID {}: {}", id, e.getMessage()))
-                    .doOnError(SecurityException.class, 
-                        e -> log.error("Authentication error retrieving statistics for load {}: {}", id, e.getMessage()))
-                    .block();
-        } catch (Exception e) {
-            if (e.getCause() instanceof IllegalArgumentException) {
-                throw new McpError(new McpSchema.JSONRPCResponse.JSONRPCError(
-                    McpSchema.ErrorCodes.INVALID_PARAMS, 
-                    "Load not found: " + e.getCause().getMessage(), 
-                    null));
-            }
-            throw handleException("Failed to retrieve load statistics", e);
-        }
+        return Mono.just(id)
+                .filter(loadId -> loadId != null && loadId > 0)
+                .switchIfEmpty(Mono.error(new McpError(new JSONRPCError(
+                        INVALID_PARAMS,
+                        "Load ID must be a positive number",
+                        null))))
+                .flatMapMany(loadId -> ReactiveSecurityContextHolder.getContext()
+                        .map(SecurityContext::getAuthentication)
+                        .flatMapMany(auth -> {
+                            var instances = discoveryClient.getInstances(loadsServiceName);
+                            if (instances == null || instances.isEmpty()) {
+                                throw new McpError(new JSONRPCError(
+                                        INTERNAL_ERROR, format("Service %s not found in discovery", loadsServiceName),
+                                        null));
+                            }
+
+                            return webClient
+                                    .get()
+                                    .uri(instances.get(0).getUri().toString() + "/loads/" + loadId + "/statistics")
+                                    .headers(h -> h.setBearerAuth(((Jwt) auth.getCredentials()).getTokenValue()))
+                                    .retrieve()
+                                    .bodyToFlux(GroupStatisticsDto.class)
+                                    .onErrorMap(WebClientResponseException.class, e -> {
+                                        if (e.getStatusCode() == UNAUTHORIZED) {
+                                            throw new McpError(new JSONRPCError(
+                                                    INVALID_REQUEST,
+                                                    "Authentication failed",
+                                                    null));
+                                        }
+                                        if (e.getStatusCode() == NOT_FOUND) {
+                                            throw new McpError(new JSONRPCError(
+                                                    INVALID_PARAMS,
+                                                    "Load not found with ID: " + loadId,
+                                                    null));
+                                        }
+                                        return e;
+                                    });
+                        }))
+                .retryWhen(Retry.backoff(3, Duration.ofSeconds(1))
+                        .filter(throwable -> !(throwable instanceof McpError)))
+                .contextWrite(ctx -> ctx.putAll(getReactiveContext()))
+                .collectList()
+                .doOnSuccess(stats -> log.debug("Retrieved {} statistics for load {}", stats.size(), id))
+                .block();
     }
 
     /**
@@ -162,40 +280,41 @@ public class LoadsTools {
     private ContextView getReactiveContext() {
         var reactiveContext = ReactiveContextHolder.reactiveContext.get();
         if (reactiveContext == null) {
-            throw new McpError(new McpSchema.JSONRPCResponse.JSONRPCError(
-                McpSchema.ErrorCodes.INTERNAL_ERROR, 
-                "No reactive context available", 
-                null));
+            throw new McpError(new JSONRPCError(
+                    INTERNAL_ERROR,
+                    "No reactive context available",
+                    null));
         }
 
         return reactiveContext;
     }
 
     /**
-     * Handles exceptions by mapping them to appropriate McpError with JSON-RPC error codes.
+     * Handles exceptions by mapping them to appropriate McpError with JSON-RPC
+     * error codes.
      * Special handling for authentication and state-related errors.
      *
      * @param message Base error message
-     * @param e Original exception
+     * @param e       Original exception
      * @return McpError with appropriate JSON-RPC error code
      */
     private McpError handleException(String message, Exception e) {
         if (e instanceof IllegalStateException) {
-            return new McpError(new McpSchema.JSONRPCResponse.JSONRPCError(
-                McpSchema.ErrorCodes.INTERNAL_ERROR,
-                e.getMessage(),
-                null));
+            return new McpError(new JSONRPCError(
+                    INTERNAL_ERROR,
+                    e.getMessage(),
+                    null));
         }
         if (e.getCause() instanceof SecurityException) {
-            return new McpError(new McpSchema.JSONRPCResponse.JSONRPCError(
-                McpSchema.ErrorCodes.INVALID_REQUEST,
-                "Authentication failed: " + e.getMessage(),
-                null));
+            return new McpError(new JSONRPCError(
+                    INVALID_REQUEST,
+                    "Authentication failed: " + e.getMessage(),
+                    null));
         }
         log.error(message, e);
-        return new McpError(new McpSchema.JSONRPCResponse.JSONRPCError(
-            McpSchema.ErrorCodes.INTERNAL_ERROR,
-            message + ": " + e.getMessage(),
-            null));
+        return new McpError(new JSONRPCError(
+                INTERNAL_ERROR,
+                message + ": " + e.getMessage(),
+                null));
     }
 }
